Cardano.Ledger.Mary.Scripts

Documentation

module Cardano.Ledger.Allegra.Scripts

Orphan instances

Crypto c ⇒ AllegraEraScript (MaryEra c)

Instance details

Methods

mkTimeStart ∷ SlotNo → NativeScript (MaryEra c)

getTimeStart ∷ NativeScript (MaryEra c) → Maybe SlotNo

mkTimeExpire ∷ SlotNo → NativeScript (MaryEra c)

getTimeExpire ∷ NativeScript (MaryEra c) → Maybe SlotNo

Crypto c ⇒ EraScript (MaryEra c)
Since Timelock scripts are a strictly backwards compatible extension of MultiSig scripts, we can use the same scriptPrefixTag tag here as we did for the ValidateScript instance in MultiSig

Instance details

Associated Types

type Script (MaryEra c) = (r ∷ Type)

type NativeScript (MaryEra c) = (r ∷ Type)

Methods

upgradeScript ∷ Script (PreviousEra (MaryEra c)) → Script (MaryEra c)

scriptPrefixTag ∷ Script (MaryEra c) → ByteString

getNativeScript ∷ Script (MaryEra c) → Maybe (NativeScript (MaryEra c))

fromNativeScript ∷ NativeScript (MaryEra c) → Script (MaryEra c)

Crypto c ⇒ ShelleyEraScript (MaryEra c)

Instance details

Methods

mkRequireSignature ∷ KeyHash 'Witness (EraCrypto (MaryEra c)) → NativeScript (MaryEra c)

getRequireSignature ∷ NativeScript (MaryEra c) → Maybe (KeyHash 'Witness (EraCrypto (MaryEra c)))

mkRequireAllOf ∷ StrictSeq (NativeScript (MaryEra c)) → NativeScript (MaryEra c)

getRequireAllOf ∷ NativeScript (MaryEra c) → Maybe (StrictSeq (NativeScript (MaryEra c)))

mkRequireAnyOf ∷ StrictSeq (NativeScript (MaryEra c)) → NativeScript (MaryEra c)

getRequireAnyOf ∷ NativeScript (MaryEra c) → Maybe (StrictSeq (NativeScript (MaryEra c)))

mkRequireMOf ∷ Int → StrictSeq (NativeScript (MaryEra c)) → NativeScript (MaryEra c)

getRequireMOf ∷ NativeScript (MaryEra c) → Maybe (Int, StrictSeq (NativeScript (MaryEra c)))