Cardano.Ledger.Mary.TxBody

Documentation

data MaryTxBody era where

Bundled Patterns

pattern MaryTxBody ∷ EraTxOut era ⇒ Set (TxIn (EraCrypto era)) → StrictSeq (TxOut era) → StrictSeq (DCert (EraCrypto era)) → Withdrawals (EraCrypto era) → Coin → ValidityInterval → StrictMaybe (Update era) → StrictMaybe (AuxiliaryDataHash (EraCrypto era)) → MultiAsset (EraCrypto era) → MaryTxBody era
A pattern to keep the newtype and the MemoBytes hidden

class AllegraEraTxBody era ⇒ MaryEraTxBody era where

Methods

mintTxBodyL ∷ Lens' (TxBody era) (MultiAsset (EraCrypto era))

mintValueTxBodyF ∷ SimpleGetter (TxBody era) (Value era)

mintedTxBodyF ∷ SimpleGetter (TxBody era) (Set (ScriptHash (EraCrypto era)))

data StrictMaybe a

Constructors

SNothing

SJust !a

data ValidityInterval

ValidityInterval is a half open interval. Closed on the bottom, open on the top. A SNothing on the bottom is negative infinity, and a SNothing on the top is positive infinity

Constructors

ValidityInterval

Fields

invalidBefore ∷ !(StrictMaybe SlotNo)

invalidHereafter ∷ !(StrictMaybe SlotNo)

Orphan instances

Crypto c ⇒ AllegraEraTxBody (MaryEra c)

Instance details

Methods

vldtTxBodyL ∷ Lens' (TxBody (MaryEra c)) ValidityInterval

Crypto c ⇒ EraTxBody (MaryEra c)

Instance details

Associated Types

type TxBody (MaryEra c) = (r ∷ Type)

Methods

mkBasicTxBody ∷ TxBody (MaryEra c)

inputsTxBodyL ∷ Lens' (TxBody (MaryEra c)) (Set (TxIn (EraCrypto (MaryEra c))))

outputsTxBodyL ∷ Lens' (TxBody (MaryEra c)) (StrictSeq (TxOut (MaryEra c)))

feeTxBodyL ∷ Lens' (TxBody (MaryEra c)) Coin

withdrawalsTxBodyL ∷ Lens' (TxBody (MaryEra c)) (Withdrawals (EraCrypto (MaryEra c)))

auxDataHashTxBodyL ∷ Lens' (TxBody (MaryEra c)) (StrictMaybe (AuxiliaryDataHash (EraCrypto (MaryEra c))))

allInputsTxBodyF ∷ SimpleGetter (TxBody (MaryEra c)) (Set (TxIn (EraCrypto (MaryEra c))))

Crypto c ⇒ MaryEraTxBody (MaryEra c)

Instance details

Methods

mintTxBodyL ∷ Lens' (TxBody (MaryEra c)) (MultiAsset (EraCrypto (MaryEra c)))

mintValueTxBodyF ∷ SimpleGetter (TxBody (MaryEra c)) (Value (MaryEra c))

mintedTxBodyF ∷ SimpleGetter (TxBody (MaryEra c)) (Set (ScriptHash (EraCrypto (MaryEra c))))

Crypto c ⇒ ShelleyEraTxBody (MaryEra c)

Instance details

Methods

ttlTxBodyL ∷ Lens' (TxBody (MaryEra c)) SlotNo

updateTxBodyL ∷ Lens' (TxBody (MaryEra c)) (StrictMaybe (Update (MaryEra c)))

updateTxBodyG ∷ SimpleGetter (TxBody (MaryEra c)) (StrictMaybe (Update (MaryEra c)))

certsTxBodyL ∷ Lens' (TxBody (MaryEra c)) (StrictSeq (DCert (EraCrypto (MaryEra c))))

certsTxBodyG ∷ SimpleGetter (TxBody (MaryEra c)) (StrictSeq (DCert (EraCrypto (MaryEra c))))