Cardano.Ledger.Mary.Value

Documentation

newtype PolicyID

Policy ID

Constructors

PolicyID

Fields

policyID ∷ ScriptHash

newtype AssetName

Asset Name

Constructors

AssetName

Fields

assetNameBytes ∷ ShortByteString

data MaryValue

The Value representing MultiAssets

Constructors

MaryValue !Coin !MultiAsset

newtype MultiAsset

The MultiAssets map

Constructors

MultiAsset (Map PolicyID (Map AssetName Integer))

insert ∷ (Integer → Integer → Integer) → PolicyID → AssetName → Integer → MultiAsset → MultiAsset

Deprecated: In favor of insertMultiAsset

insert comb policy asset n v, if comb = old new -> old, the integer in the MultiAsset is prefered over n if comb = old new -> new, then n is prefered over the integer in the MultiAsset if (comb old new) == 0, then that value should not be stored in the MultiAsset

insertMultiAsset ∷ (Integer → Integer → Integer) → PolicyID → AssetName → Integer → MultiAsset → MultiAsset

insertMultiAsset comb policy asset n v, if comb = old new -> old, the integer in the MultiAsset is prefered over n if comb = old new -> new, then n is prefered over the integer in the MultiAsset if (comb old new) == 0, then that value should not be stored in the MultiAsset

lookup ∷ PolicyID → AssetName → MaryValue → Integer

Deprecated: In favor of lookupMultiAsset

lookupMultiAsset ∷ PolicyID → AssetName → MaryValue → Integer

multiAssetFromList ∷ [(PolicyID, AssetName, Integer)] → MultiAsset

Rather than using prune to remove 0 assets, when can avoid adding them in the first place by using valueFromList to construct a MultiAsset

policies ∷ MultiAsset → Set PolicyID

Extract the set of policies in the Value.

This function is equivalent to computing the support of the value in the spec.

mapMaybeMultiAsset ∷ (PolicyID → AssetName → Integer → Maybe Integer) → MultiAsset → MultiAsset

Map a function over each multi asset value while optionally filtering values out. Canonical form is preserved.

filterMultiAsset

Arguments

∷ (PolicyID → AssetName → Integer → Bool)
Predicate that needs to return True whenever an asset should be retained.

→ MultiAsset

→ MultiAsset

Filter multi assets. Canonical form is preserved.

pruneZeroMultiAsset ∷ MultiAsset → MultiAsset

Remove all assets with that have zero amount specified

representationSize ∷ [(PolicyID, AssetName, Integer)] → Int

showValue ∷ MaryValue → String

Display a MaryValue as a String, one token per line

flattenMultiAsset ∷ MultiAsset → [(PolicyID, AssetName, Integer)]

valueFromList ∷ Coin → [(PolicyID, AssetName, Integer)] → MaryValue

data CompactValue

Constructors

CompactValueAdaOnly !(CompactForm Coin)

CompactValueMultiAsset !(CompactForm Coin) !Word32 !ShortByteString

data family CompactForm a

isMultiAssetSmallEnough ∷ MultiAsset → Bool

Unlike representationSize, this function cheaply checks whether any offset within a MultiAsset compact representation is likely to overflow Word16.

compact form inequality: 8n (Word64) + 2n (Word16) + 2n (Word16) + 28p (policy ids) + sum of lengths of unique asset names <= 65535 maximum size for the asset name is 32 bytes, so: 8n + 2n + 2n + 28p + 32n <= 65535 where: n = total number of assets, p = number of unique policy ids

assetNameToTextAsHex ∷ AssetName → Text

Deprecated

prune ∷ Map PolicyID (Map AssetName Integer) → Map PolicyID (Map AssetName Integer)

Deprecated: In favor of pruneZeroMultiAsset

Remove 0 assets from a MultiAsset