Cardano.Ledger.Mary.Scripts

Documentation

translateTimelock ∷ (Era era1, Era era2, EraCrypto era1 ~ EraCrypto era2) ⇒ Timelock era1 → Timelock era2

This function deconstructs and then reconstructs the timelock script to prove the compiler that we can arbirarily switch out the eras as long as the cryptos for both eras are the same.

data Timelock era

Orphan instances

Crypto c ⇒ EraScript (MaryEra c)
Since Timelock scripts are a strictly backwards compatible extension of MultiSig scripts, we can use the same scriptPrefixTag tag here as we did for the ValidateScript instance in MultiSig

Instance details

Associated Types

type Script (MaryEra c) = (r ∷ Type)

type NativeScript (MaryEra c) = (r ∷ Type)

Methods

upgradeScript ∷ Script (PreviousEra (MaryEra c)) → Script (MaryEra c)

scriptPrefixTag ∷ Script (MaryEra c) → ByteString

getNativeScript ∷ Script (MaryEra c) → Maybe (NativeScript (MaryEra c))

fromNativeScript ∷ NativeScript (MaryEra c) → Script (MaryEra c)