Cardano.Ledger.Mary.Scripts

Documentation

module Cardano.Ledger.Allegra.Scripts

Orphan instances

Crypto c ⇒ EraScript (MaryEra c)
Since Timelock scripts are a strictly backwards compatible extension of MultiSig scripts, we can use the same scriptPrefixTag tag here as we did for the ValidateScript instance in MultiSig

Instance details

Associated Types

type Script (MaryEra c) = (r ∷ Type)

Methods

scriptPrefixTag ∷ Script (MaryEra c) → ByteString

hashScript ∷ Script (MaryEra c) → ScriptHash (EraCrypto (MaryEra c))

phaseScript ∷ ∀ (phase ∷ Phase). PhaseRep phase → Script (MaryEra c) → Maybe (PhasedScript phase (MaryEra c))

isNativeScript ∷ Script (MaryEra c) → Bool