Cardano.Ledger.Mary.UTxO

Documentation

getConsumedMaryValue ∷ (MaryEraTxBody era, Value era ~ MaryValue (EraCrypto era)) ⇒ PParams era → (Credential 'Staking (EraCrypto era) → Maybe Coin) → (Credential 'DRepRole (EraCrypto era) → Maybe Coin) → UTxO era → TxBody era → MaryValue (EraCrypto era)

Calculate the value consumed by the transation.

This differs from the corresponding Shelley function coinConsumed since it works on Value and it also considers the "mint" field which creates or destroys non-Ada tokens.

Note that this is slightly confusing, since it also covers non-Ada assets _created_ by the transaction, depending on the sign of the quantities in the mint field.

getProducedMaryValue

Arguments

∷ (MaryEraTxBody era, Value era ~ MaryValue (EraCrypto era))

⇒ PParams era

→ (KeyHash 'StakePool (EraCrypto era) → Bool)
Check whether a pool with a supplied PoolStakeId is already registered.

→ TxBody era

→ MaryValue (EraCrypto era)

Orphan instances

Crypto c ⇒ EraUTxO (MaryEra c)

Instance details

Associated Types

type ScriptsNeeded (MaryEra c) = (r ∷ Type)

Methods

getConsumedValue ∷ PParams (MaryEra c) → (Credential 'Staking (EraCrypto (MaryEra c)) → Maybe Coin) → (Credential 'DRepRole (EraCrypto (MaryEra c)) → Maybe Coin) → UTxO (MaryEra c) → TxBody (MaryEra c) → Value (MaryEra c)

getProducedValue ∷ PParams (MaryEra c) → (KeyHash 'StakePool (EraCrypto (MaryEra c)) → Bool) → TxBody (MaryEra c) → Value (MaryEra c)

getScriptsProvided ∷ UTxO (MaryEra c) → Tx (MaryEra c) → ScriptsProvided (MaryEra c)

getScriptsNeeded ∷ UTxO (MaryEra c) → TxBody (MaryEra c) → ScriptsNeeded (MaryEra c)

getScriptsHashesNeeded ∷ ScriptsNeeded (MaryEra c) → Set (ScriptHash (EraCrypto (MaryEra c)))

getWitsVKeyNeeded ∷ CertState (MaryEra c) → UTxO (MaryEra c) → TxBody (MaryEra c) → Set (KeyHash 'Witness (EraCrypto (MaryEra c)))

getMinFeeTxUtxo ∷ PParams (MaryEra c) → Tx (MaryEra c) → UTxO (MaryEra c) → Coin