Cardano.Ledger.Mary.Value

Documentation

newtype PolicyID c

Policy ID

Constructors

PolicyID

Fields

policyID ∷ ScriptHash c

newtype AssetName

Asset Name

Constructors

AssetName

Fields

assetName ∷ ShortByteString

data MaryValue c

The Value representing MultiAssets

Constructors

MaryValue !Integer !(MultiAsset c)

newtype MultiAsset c

The MultiAssets map

Constructors

MultiAsset (Map (PolicyID c) (Map AssetName Integer))

insert ∷ (Integer → Integer → Integer) → PolicyID c → AssetName → Integer → MultiAsset c → MultiAsset c

Deprecated: In favor of insertMultiAsset

insert comb policy asset n v, if comb = old new -> old, the integer in the MultiAsset is prefered over n if comb = old new -> new, then n is prefered over the integer in the MultiAsset if (comb old new) == 0, then that value should not be stored in the MultiAsset

insertMultiAsset ∷ (Integer → Integer → Integer) → PolicyID c → AssetName → Integer → MultiAsset c → MultiAsset c

insertMultiAsset comb policy asset n v, if comb = old new -> old, the integer in the MultiAsset is prefered over n if comb = old new -> new, then n is prefered over the integer in the MultiAsset if (comb old new) == 0, then that value should not be stored in the MultiAsset

lookup ∷ PolicyID c → AssetName → MaryValue c → Integer

Deprecated: In favor of lookupMultiAsset

lookupMultiAsset ∷ PolicyID c → AssetName → MaryValue c → Integer

multiAssetFromList ∷ [(PolicyID era, AssetName, Integer)] → MultiAsset era

Rather than using prune to remove 0 assets, when can avoid adding them in the first place by using valueFromList to construct a MultiAsset

policies ∷ MultiAsset c → Set (PolicyID c)

Extract the set of policies in the Value.

This function is equivalent to computing the support of the value in the spec.

prune ∷ Map (PolicyID c) (Map AssetName Integer) → Map (PolicyID c) (Map AssetName Integer)

Remove 0 assets from a map

representationSize ∷ ∀ c. Crypto c ⇒ [(PolicyID c, AssetName, Integer)] → Int

showValue ∷ MaryValue c → String

Display a MaryValue as a String, one token per line

flattenMultiAsset ∷ MultiAsset c → [(PolicyID c, AssetName, Integer)]

valueFromList ∷ Integer → [(PolicyID era, AssetName, Integer)] → MaryValue era

class ToExpr a where

Minimal complete definition

Nothing

Methods

toExpr ∷ a → Expr

listToExpr ∷ [a] → Expr

data CompactValue c

Constructors

CompactValueAdaOnly !(CompactForm Coin)

CompactValueMultiAsset !(CompactForm Coin) !Word32 !ShortByteString

isMultiAssetSmallEnough ∷ MultiAsset c → Bool

Unlike representationSize, this function cheaply checks whether any offset within a MultiAsset compact representation is likely to overflow Word16.

compact form inequality: 8n (Word64) + 2n (Word16) + 2n (Word16) + 28p (policy ids) + sum of lengths of unique asset names <= 65535 maximum size for the asset name is 32 bytes, so: 8n + 2n + 2n + 28p + 32n <= 65535 where: n = total number of assets, p = number of unique policy ids