Test.Cardano.Ledger.Mary.Arbitrary

Documentation

genEmptyMultiAsset ∷ Crypto c ⇒ Gen (MultiAsset c)

genMaryValue ∷ Gen (MultiAsset c) → Gen (MaryValue c)

genMultiAsset ∷ ∀ c. Crypto c ⇒ Gen Integer → Gen (MultiAsset c)

genMultiAssetToFail ∷ ∀ c. Crypto c ⇒ Bool → Gen (MultiAsset c)

genMultiAssetZero ∷ Crypto c ⇒ Gen (MultiAsset c)

For tests, because insertMultiAsset called through genMultiAsset filters out zero values

genPositiveInt ∷ Gen Int

Better generator for a positive Int that explores more values

genNegativeInt ∷ Gen Int

Better generator for a Negative Int that explores more values

genNonNegativeInt ∷ Gen Int

Better generator for a Non-Negative Int that explores more values

Orphan instances

Arbitrary AssetName

Instance details

Methods

arbitrary ∷ Gen AssetName

shrink ∷ AssetName → [AssetName]

Crypto c ⇒ Arbitrary (CompactForm (MaryValue c))

Instance details

Methods

arbitrary ∷ Gen (CompactForm (MaryValue c))

shrink ∷ CompactForm (MaryValue c) → [CompactForm (MaryValue c)]

(EraTxOut era, EraTxCert era, Era era, Arbitrary (TxOut era), Arbitrary (PParamsHKD StrictMaybe era), Arbitrary (TxCert era)) ⇒ Arbitrary (MaryTxBody era)

Instance details

Methods

arbitrary ∷ Gen (MaryTxBody era)

shrink ∷ MaryTxBody era → [MaryTxBody era]

Crypto c ⇒ Arbitrary (MaryValue c)

Instance details

Methods

arbitrary ∷ Gen (MaryValue c)

shrink ∷ MaryValue c → [MaryValue c]

Crypto c ⇒ Arbitrary (MultiAsset c)

Instance details

Methods

arbitrary ∷ Gen (MultiAsset c)

shrink ∷ MultiAsset c → [MultiAsset c]

Crypto c ⇒ Arbitrary (PolicyID c)

Instance details

Methods

arbitrary ∷ Gen (PolicyID c)

shrink ∷ PolicyID c → [PolicyID c]